+++ iOS/ActionComponentExplorer/ActionComponentExplorer/ViewController.designer.cs
@@ -13,13 +13,45 @@ namespace ActionComponentExplorer
 	partial class ViewController
 	{
 		[Outlet]
+		UIKit.UIScrollView ScrollView { get; set; }
+
+		[Outlet]
 		ActionComponents.ACTileController tileController { get; set; }
+
+		[Outlet]
+		ActionComponents.ACTray TrayOne { get; set; }
+
+		[Outlet]
+		ActionComponents.ACTray TrayThree { get; set; }
+
+		[Outlet]
+		ActionComponents.ACTray TrayTwo { get; set; }
 		
 		void ReleaseDesignerOutlets ()
 		{
 			if (tileController != null) {
 				tileController.Dispose ();
 				tileController = null;
+			}
+
+			if (TrayOne != null) {
+				TrayOne.Dispose ();
+				TrayOne = null;
+			}
+
+			if (ScrollView != null) {
+				ScrollView.Dispose ();
+				ScrollView = null;
+			}
+
+			if (TrayTwo != null) {
+				TrayTwo.Dispose ();
+				TrayTwo = null;
+			}
+
+			if (TrayThree != null) {
+				TrayThree.Dispose ();
+				TrayThree = null;
 			}
 		}
 	}